+++ digitalforms-api/src/main/java/ca/bc/gov/open/pssg/rsbc/digitalforms/controller/PaymentServiceController.java
@@ -3,15 +3,16 @@ package ca.bc.gov.open.pssg.rsbc.digitalforms.controller;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
+import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RestController;
 
 import ca.bc.gov.open.pssg.rsbc.digitalforms.model.JSONResponse;
+import ca.bc.gov.open.pssg.rsbc.digitalforms.model.PaymentStatusResponse;
 import ca.bc.gov.open.pssg.rsbc.digitalforms.model.PaymentTransRequest;
 import ca.bc.gov.open.pssg.rsbc.digitalforms.service.PaymentService;
-import ca.bc.gov.open.pssg.rsbc.digitalforms.service.PaymentServiceImpl;
 import io.swagger.annotations.Api;
 import io.swagger.annotations.ApiOperation;
 import io.swagger.annotations.ApiResponse;
@@ -30,23 +31,39 @@ public class PaymentServiceController {
 	
 	// Provides generic type class defs for Swagger 2. 
 	private class ReviewPaidSwaggerResponse extends JSONResponse<Boolean>{}
+	private class PaymentStatusSwaggerResponse extends JSONResponse<PaymentStatusResponse> {}
 	
 	@Autowired 
 	private PaymentService paymentService; 
 	
-	public PaymentServiceController(PaymentServiceImpl paymentService) {
+	public PaymentServiceController(PaymentService paymentService) {
 		this.paymentService = paymentService;
 	}
 	
 	@ApiOperation(value = "Set Prohibition Review Paid", response = ReviewPaidSwaggerResponse.class) 
 	@ApiResponses(value = {@ApiResponse(code = 200, message = "Success", response = ReviewPaidSwaggerResponse.class)})
 	@PostMapping(path ="/{noticeNumber}/payment",
-	consumes = "application/json",
-	produces = "application/json"
+		consumes = "application/json",
+		produces = "application/json"
 	)	
-	public ResponseEntity<JSONResponse<Boolean>> setReviewPaid(@PathVariable (value="noticeNumber", required=true) Long noticeNumber, @RequestBody (required=true) PaymentTransRequest paymentInfo)  {
+	public ResponseEntity<JSONResponse<Boolean>> setReviewPaid(
+			@PathVariable (value="noticeNumber", required=true) Long noticeNumber, 
+			@RequestBody (required=true) PaymentTransRequest paymentInfo)  {
 	    boolean data = paymentService.setReviewPaid(noticeNumber, paymentInfo);
 	    JSONResponse<Boolean> resp = new JSONResponse<>(data);
 	    return new ResponseEntity<>(resp, HttpStatus.OK);
+	}
+	
+	@ApiOperation(value = "Get Payment Status", response = PaymentStatusSwaggerResponse.class)
+	@ApiResponses(value = {
+			@ApiResponse(code = 200, message = "Success", response = PaymentStatusSwaggerResponse.class) })
+	@GetMapping(value = "{noticeNumber}/payment/status", 
+			produces = "application/json"
+    )
+	public ResponseEntity<JSONResponse<PaymentStatusResponse>> paymentStatusGet(
+			@PathVariable(value = "noticeNumber", required = true) Long noticeNumber) {
+		PaymentStatusResponse data = paymentService.getReviewPaymentStatus(noticeNumber);
+		JSONResponse<PaymentStatusResponse> resp = new JSONResponse<>(data);
+		return new ResponseEntity<>(resp, HttpStatus.OK);
 	}
 }
